requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
pymupdf==1.22.0
pillow==10.0.1
pytesseract==0.3.10
transformers==4.35.0
torch==2.2.0
